plugin pcg-interactive-blocks
This commit introduces the following blocks

## `pcg-dice-roller`

This block is a container for the single die. When placed in the editor, it generates a "d4", "d6", "d8", "d10", "d12" and "d20" `single-die` blocks as its children. On the frontend, it displays the total which is a sum of the `currentValue` from context from each `single-die` block as its "rolled".

## `pcg-single-die`

This block is a die that can be rolled on the frontend. It "rolls" when clicked and will "land" on a number. When a die lands on its largest possible value, it "explodes".

These blocks leverage the WordPress Interactivity API.